Rename HimmeslrichtungUtil into QuUtils.ts
