+++ frontend/src/util/QuUtils.ts
